VAC 패턴 적용(components/Footer/) #9
